requirements.txt
numpy
matplotlib
#matplotlib.pyplot
pandas
IPython
cython
pykka
pillow
psutil